Rename CANON->EPIC, EPIC->NPEC, Return->ERC to match paper convention
