+++ runners/comparison/hardcoded_figs.sh
