+++ runners/comparison/hardcoded_npec.sh
